move divisor tests into seperate file
